Modified ordering of gates when printing circuits
quantum_circuit.cc: reversed the order of elements in the vector
that holds the gates, as strings, of the circuit.
io_helper.py: Fixed bug in the smart_print function in the case
of quantum circuits. The |Ψ> was inserted to emphasize the order
in which the operators are applied. This needs approval.